:recycle:Refactor : 글작성 페이지와 수정 페이지 분리 및 코드 컴포넌트화 (#126)
