+++ app/(main)/mydiary/components/tabs.tsx
