+++ app/(main)/mydiary/write-diary-btn.tsx
@@ -4,13 +4,13 @@ import { Pencil } from "lucide-react"
 import { useRouter } from "next/navigation"
 
 export const WriteDiaryBtn = () => {
-  const router = useRouter()
+  const { push } = useRouter()
   return (
     <div className="fixed bottom-20 right-4 z-10 bg-mainColor p-3 rounded-full">
       <div
         className=" boxshadow"
         onClick={() => {
-          router.push(`/mydiary/write-diary`)
+          push("/mydiary/new/write-diary")
         }}>
         <Pencil size={28} />
       </div>
