+++ app/(onboarding)/welcome/page.tsx
@@ -34,7 +34,7 @@ export default function WelcomePage() {
       <Button
         type="button"
         className="fixed bottom-6 w-5/6 max-w-96 left-1/2 -translate-x-1/2 bg-mainColor text-black font-bold "
-        onClick={() => router.push(`/mydiary/write-diary`)}>
+        onClick={() => router.push(`/mydiary/new/write-diary`)}>
         일기 작성하기
       </Button>
     </div>
